Agrega custom useNearScreen useLocalStorage hooks
Funcionalidades convertidas en custom hooks para mejor reusabilidad
