+++ src/components/PhotoCard/index.js
@@ -1,41 +1,15 @@
 import React, { useEffect, useRef, useState } from 'react'
+import {useLocalStorage} from '../../hooks/useLocalStorage'
+import { useNearScreen } from '../../hooks/useNearScreen'
 import { ImgWrapper, Image, Button, Article } from './styles'
 import { MdFavoriteBorder, MdFavorite } from 'react-icons/md'
 
 const DEFAULT_IMAGE = 'https://res.cloudinary.com/midudev/image/upload/w_300/q_80/v1560262103/dogs.png'
 
 export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
-  const ref = useRef(null)
-  const [show, setShow] = useState(false)
+  const [show,element]=useNearScreen()
   const key = `like-${id}`
-  const [liked, setLiked] = useState(() => {
-    try {
-      const like = window.localStorage.getItem(key)
-      return JSON.parse(like)
-    } catch (e) {
-      console.log(e)
-      return false
-    }
-  })
-
-  console.log(liked)
-
-  useEffect(() => {
-    Promise.resolve(
-      typeof window.IntersectionObserver !== 'undefined'
-        ? window.IntersectionObserver
-        : import('intersection-observer')
-    ).then(() => {
-      const observer = new window.IntersectionObserver((entries) => {
-        const { isIntersecting } = entries[0]
-        if (isIntersecting) {
-          setShow(true)
-          observer.disconnect()
-        }
-      })
-      observer.observe(ref.current)
-    })
-  }, [ref])
+  const [liked, setLiked]=useLocalStorage(key, false)
 
   const Icon = liked ? MdFavorite : MdFavoriteBorder
 
@@ -49,7 +23,7 @@ export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
   }
 
   return (
-    <Article ref={ref}>
+    <Article ref={element}>
       {
         show &&
           <>
@@ -58,7 +32,7 @@ export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
                 <Image src={src} />
               </ImgWrapper>
             </a>
-            <Button onClick={() => setLocalStorage(!liked)}>
+            <Button onClick={() => setLiked(!liked)}>
               <Icon size='32px' /> {likes} likes!
             </Button>
           </>
